third_party/WebKit/LayoutTests/inspector/file-reader-with-network-panel.html

<html>
<head>
<script src="../http/tests/inspector/inspector-test.js"></script>
<script src="../http/tests/inspector/network-test.js"></script>
<script>
function readBlob()
{
    var reader = new FileReader();
    reader.onloadend = function () {
        console.log('done');
    };
    reader.readAsArrayBuffer(new Blob([ 'test' ]));
}

function test()
{
    InspectorTest.addConsoleSniffer(messageAdded);
    InspectorTest.evaluateInPage("readBlob();");

    function messageAdded(payload)
    {
        var requests = InspectorTest.networkRequests();
        InspectorTest.addResult("requests in the network panel: " + requests.length);
        InspectorTest.assertTrue(requests.length == 0, "Blob load request to the browser is shown in the network panel.");
        InspectorTest.completeTest();
    }
}

</script>
</head>
<body onload="runTest()">
<p>Tests that FileReader's Blob request isn't shown in network panel.</p>
</body>
</html>
