third_party/WebKit/LayoutTests/inspector/file-reader-with-network-panel-expected.txt

CONSOLE MESSAGE: line 10: done
Tests that FileReader's Blob request isn't shown in network panel.

requests in the network panel: 0
